Benchmarks/google/requirements.txt
# Function dependencies, for example:
# package>=version
speedtest-cli